Regroup the translation of spans and attributes and such
