+++ charon/src/bin/charon-driver/translate/mod.rs
@@ -8,5 +8,6 @@ pub mod translate_ctx;
 pub mod translate_functions;
 pub mod translate_generics;
 pub mod translate_items;
+pub mod translate_meta;
 pub mod translate_predicates;
 pub mod translate_types;
